requirements.txt
coincurve==21.0.0
eth-account==0.13.7
eth-hash==0.7.1
mnemonic==0.21
python-dotenv==1.2.1
pycryptodome==3.23.0
web3==7.14.1
fastapi
uvicorn
python-multipart
